Refactor: Remove Modes.css and update SimplePanel to use its own css file
- Deleted Modes.css as it is no longer needed.
- Updated SimplePanel.tsx to remove import of Modes.css.
- Removed explanationMode from settings schema and related functions.
- Cleaned up store.ts by removing unused ExplanationMode type.
- Added new AdvancedPanel.css with styles for advanced panel components.
- Removed unused css styles
