+++ src/components/panels/AdvancedPanel.css
@@ -1,81 +1,3 @@
-/* =====================
-   SIMPLE
-   ===================== */
-
-.simple-panel {
-  display: flex;
-  flex-direction: column;
-  justify-content: center;
-  gap: 1rem;
-  height: 80%;
-  padding: 0 1.25rem;
-}
-
-.simple-row {
-  display: flex;
-  align-items: center;
-  justify-content: center;
-  gap: 1.25rem;
-}
-
-.simple-group {
-  display: flex;
-  align-items: center;
-  gap: 0.5rem;
-}
-
-.simple-label {
-  font-family: 'DMSans', sans-serif;
-  font-size: 1rem;
-  font-weight: var(--font-weight-md);
-  letter-spacing: -0.02em;
-  color: var(--text-muted);
-  white-space: nowrap;
-}
-
-.simple-input-box {
-  display: flex;
-  align-items: center;
-  gap: 0.3125rem;
-  background: var(--bg-elevated);
-  border: 1px solid var(--border-subtle);
-  border-radius: var(--radius-sm);
-  padding: 2px 0.75rem;
-  min-width: 6.25rem;
-  transition: border-color 0.15s ease;
-}
-
-.simple-input-box:hover {
-  border-color: var(--text-dim);
-}
-
-.simple-number {
-  font-family: 'DMSans', sans-serif;
-  font-size: 0.9375rem;
-  font-weight: var(--font-weight-md);
-  color: var(--text-primary);
-}
-
-.simple-unit {
-  font-family: 'DMSans', sans-serif;
-  font-size: 0.875rem;
-  font-weight: var(--font-weight-hv);
-  letter-spacing: -0.02em;
-  color: var(--text-dim);
-}
-
-.simple-hotkey {
-  min-width: 6.25rem;
-  cursor: text;
-}
-
-.simple-hotkey-text {
-  font-family: 'DMSans', sans-serif;
-  font-size: 0.9375rem;
-  font-weight: var(--font-weight-md);
-  color: var(--text-primary);
-  cursor: pointer;
-}
 
 .simple-seg-group {
   display: flex;
@@ -115,9 +37,6 @@
   outline: none;
   box-shadow: none;
 }
-/* =====================
-   ADVANCED
-   ===================== */
 
 input[type="number"]::-webkit-inner-spin-button,
 input[type="number"]::-webkit-outer-spin-button {
@@ -131,116 +50,8 @@ input[type="number"]::-webkit-outer-spin-button {
   height: 100%;
 }
 
-.advanced-panel-compact .advanced-columns {
-  gap: 0.5rem;
-}
-
-.advanced-panel-compact .sectioncontainer {
-  padding: 0.5rem 0.625rem;
-  margin-bottom: 0;
-}
-
-.advanced-panel-compact .adv-card-title {
-  font-size: 0.8125rem;
-}
-
-.advanced-panel-compact .adv-label {
-  font-size: 0.8125rem;
-}
-
-.advanced-panel-compact .adv-row {
-  gap: 0.375rem;
-}
-
-.advanced-panel-compact .adv-card-header {
-  align-items: flex-start;
-}
-
-.advanced-panel-compact .adv-card-body,
-.advanced-panel-compact .adv-limit-block {
-  gap: 0.375rem;
-}
-
-.advanced-panel-compact .adv-numbox-sm {
-  padding: 2px 0.4375rem;
-}
-
-.advanced-panel-compact .adv-card-divider {
-  margin: 0.375rem 0;
-}
-
-.advanced-panel-compact .adv-corner-grid {
-  gap: 0.3125rem;
-}
-
-.advanced-panel-compact .adv-corner-box {
-  padding: 0.1875rem 0.4375rem;
-}
-
-.advanced-panel-compact .adv-pick-btn {
-  font-size: 0.6875rem;
-  padding: 0.25rem 0.5rem;
-}
-
-.advanced-panel-compact .adv-pick-btn-inline {
-  width: auto;
-  min-width: 5.25rem;
-}
-
-.advanced-panel-compact .simple-seg-btn {
-  font-size: 0.875rem;
-  padding: 0.1875rem 0.5rem;
-}
-
-.advanced-panel-compact .adv-toggle-btn {
-  min-width: 1.9375rem;
-  font-size: 0.6875rem;
-}
-
 .adv-toggle-btn {
   text-transform: uppercase;
-}
-
-.advanced-panel-compact .adv-textbox {
-  min-width: 0;
-  flex: 1;
-}
-
-.advanced-panel-compact .adv-textbox-text {
-  width: 100% !important;
-}
-
-.adv-compact-stack {
-  display: flex;
-  flex-direction: column;
-  gap: 0.5rem;
-  height: 100%;
-}
-
-.adv-compact-top {
-  display: grid;
-  grid-template-columns: minmax(0, 2fr) minmax(11.25rem, 1fr);
-  gap: 0.5rem;
-}
-
-.adv-compact-grid {
-  display: flex;
-  flex-direction: column;
-  gap: 0.5rem;
-}
-
-.adv-compact-card {
-  min-width: 0;
-}
-
-.adv-compact-card-wide {
-  grid-column: 1 / -1;
-}
-
-.adv-compact-three-grid {
-  display: grid;
-  grid-template-columns: repeat(3, minmax(0, 1fr));
-  gap: 0.5rem;
 }
 
 .advanced-columns {
@@ -254,20 +65,6 @@ input[type="number"]::-webkit-outer-spin-button {
   display: flex;
   flex-direction: column;
   flex: 1;
-}
-
-.adv-compact-card-grid {
-  display: grid;
-  grid-template-columns: repeat(2, minmax(0, 1fr));
-  gap: 0.5rem;
-}
-
-.adv-compact-mini-card {
-  min-width: 0;
-}
-
-.adv-compact-wide-card {
-  grid-column: 1 / -1;
 }
 
 /* Card */
@@ -310,59 +107,6 @@ input[type="number"]::-webkit-outer-spin-button {
   align-items: center;
   gap: 8px;
   flex-wrap: wrap;
-}
-
-.adv-card-body,
-.adv-limit-block {
-  display: flex;
-  flex-direction: column;
-  gap: 0.5rem;
-}
-
-.adv-inline-controls {
-  display: flex;
-  align-items: center;
-  gap: 0.5rem;
-  flex-wrap: wrap;
-}
-
-.adv-inline-controls-start {
-  justify-content: flex-start;
-}
-
-.adv-feature-body {
-  display: flex;
-  align-items: flex-start;
-  gap: 0.5rem;
-}
-
-.adv-feature-body-compact {
-  justify-content: flex-start;
-}
-
-.adv-position-controls {
-  display: flex;
-  flex-direction: column;
-  align-items: center;
-  gap: 0.25rem;
-}
-
-.adv-position-controls-compact {
-  align-items: flex-start;
-}
-
-.adv-position-row {
-  display: flex;
-  gap: 0.5rem;
-  align-items: center;
-}
-
-.adv-position-inline {
-  align-items: center;
-}
-
-.advanced-panel-compact .adv-position-inline {
-  gap: 0.375rem;
 }
 
 .adv-axis-label {
@@ -419,14 +163,6 @@ input[type="number"]::-webkit-outer-spin-button {
   white-space: nowrap;
 }
 
-.adv-numbox {
-  display: flex;
-  align-items: center;
-  background: var(--bg-elevated);
-  border: 1px solid var(--border-subtle);
-  border-radius: var(--radius-md);
-  padding: 2px 0.625rem;
-}
 
 .adv-numbox-sm {
   display: flex;
@@ -436,13 +172,6 @@ input[type="number"]::-webkit-outer-spin-button {
   border: 1px solid var(--border-subtle);
   border-radius: var(--radius-sm);
   padding:0.1875rem 0.375rem;
-}
-
-.adv-number {
-  font-family: var(--font-family-base);
-  font-size: 14px;
-  font-weight: var(--font-weight-hv);
-  color: var(--text-primary);
 }
 
 .adv-number-sm {
@@ -769,8 +498,3 @@ input[type="number"]::-webkit-outer-spin-button {
   cursor: not-allowed;
   pointer-events: none;
 }
-
-.advanced-panel-compact .adv-sequence-item {
-  grid-template-columns: 24px minmax(0, auto) minmax(0, auto) 1fr;
-}
-
